# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AgencyOS API
  description: Backend for the AI Outreach Agency Operating System.
  version: 0.1.0
paths:
...
  /api/v1/automation/status:
    get:
      tags:
      - automation
      summary: Get automation status
      description: 'Return the current automation enabled status and pause metadata.


        Admin-only endpoint that returns whether automation is running and any

        pause details (who paused it, when, and why).'
      operationId: get_automation_status_api_v1_automation_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingStatusResponse'
      security:
      - HTTPBearer: []
  /api/v1/automation/pause:
    post:
      tags:
      - automation
      summary: Pause automation
      description: 'Pause all automation (operator-only).


        Sets ``automation.enabled`` to false, stores pause metadata, and writes an

        ``activity_logs`` entry with ``AUTOMATION_PAUSED``. All execution phases,

        queue, and schedule dispatch will be blocked until this endpoint is called

        with ``/resume``.'
      operationId: pause_automation_api_v1_automation_pause_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingPauseRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/automation/resume:
    post:
      tags:
      - automation
      summary: Resume automation
      description: 'Resume automation after a pause (operator-only).


        Clears the pause metadata and sets ``automation.enabled`` back to true.

        Writes an ``activity_logs`` entry with ``AUTOMATION_RESUMED``.'
      operationId: resume_automation_api_v1_automation_resume_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingResumeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SystemSettingPauseRequest:
      properties:
        reason:
          type: string
          title: Reason
      type: object
      required:
      - reason
      title: SystemSettingPauseRequest
    SystemSettingResumeRequest:
      properties: {}
      type: object
      title: SystemSettingResumeRequest
    SystemSettingStatusResponse:
      properties:
        enabled:
          type: boolean
          title: Enabled
        paused_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Paused By
        paused_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Paused At
        paused_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Paused Reason
      type: object
      required:
      - enabled
      title: SystemSettingStatusResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer